+++ layers/Engine/packages/graphql-parser/tests/Spec/Parser/ParserTest.php
@@ -172,6 +172,8 @@ GRAPHQL;
             ['{ test (a: "asd", b: <basd>) { id }'],
             ['{ test (asd: [..., asd]) { id } }'],
             ['{ test (asd: { "a": 4, "m": null, "asd": false  "b": 5, "c" : { a }}) { id } }'],
+            ['{ test (a: """asd") { id }'],
+            ['{ test (a: "asd""") { id }'],
             ['asdasd'],
             ['mutation { test(asd: ... ){ ...,asd, asd } }'],
             ['mutation { test{ . test on Test { id } } }'],
@@ -846,6 +848,64 @@ GRAPHQL;
                     ]
                 ),
                 'query { user(id: 10, name: "max", float: 123.123) { id name } }',
+            ],
+            // Block Strings
+            [
+                '{ user (id: 10, name: """max""", float: 123.123 ) { id, name } }',
+                new Document(
+                    [
+                        new QueryOperation('', [], [], [
+                            new RelationalField(
+                                'user',
+                                null,
+                                [
+                                    new Argument('id', new Literal(10, new Location(1, 13)), new Location(1, 9)),
+                                    new Argument('name', new Literal('max', new Location(1, 26)), new Location(1, 17)),
+                                    new Argument('float', new Literal(123.123, new Location(1, 41)), new Location(1, 34)),
+                                ],
+                                [
+                                    new LeafField('id', null, [], [], new Location(1, 53)),
+                                    new LeafField('name', null, [], [], new Location(1, 57)),
+                                ],
+                                [],
+                                new Location(1, 3)
+                            ),
+                        ], new Location(1, 1))
+                    ]
+                ),
+                'query { user(id: 10, name: "max", float: 123.123) { id name } }',
+            ],
+            // Block Strings with newlines
+            [
+                '{ user (id: 10, name: """
+                    max
+                """, float: 123.123 ) { id, name } }',
+                new Document(
+                    [
+                        new QueryOperation('', [], [], [
+                            new RelationalField(
+                                'user',
+                                null,
+                                [
+                                    new Argument('id', new Literal(10, new Location(1, 13)), new Location(1, 9)),
+                                    new Argument('name', new Literal('
+                    max
+                ', new Location(1, 26)), new Location(1, 17)),
+                                    new Argument('float', new Literal(123.123, new Location(1, 79)), new Location(1, 72)),
+                                ],
+                                [
+                                    new LeafField('id', null, [], [], new Location(1, 91)),
+                                    new LeafField('name', null, [], [], new Location(1, 95)),
+                                ],
+                                [],
+                                new Location(1, 3)
+                            ),
+                        ], new Location(1, 1))
+                    ]
+                ),
+                'query { user(id: 10, name: "
+                    max
+                ", float: 123.123) { id name } }',
             ],
             [
                 '{ allUsers : users ( id: [ 1, 2, 3] ) { id } }',