+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/CHANGELOG.md
@@ -180,6 +180,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
   - Support fragment spread on unions
   - Variables are input types
   - Queried fields are unambiguous
+- Support block string characters
 - Query schema extensions via introspection
   - Implemented extension `isSensitiveDataElement`
 - Performance improvement: Avoid regenerating the container when the schema is modified